Move gameplay logic into a service
